chore: moved components from the registration screen to the 'tela-cadastro' folder for better organization
