+++ frontend-orange/src/App.jsx
@@ -1,6 +1,6 @@
 import RegistrationPage from '../src/pages/registration-page';
 import TelaLogin from '../src/pages/TelaLogin';
-import MainHeader from '../src/components/main-header';
+import MainHeader from '../src/components/main-header/Header';
 import './main.css';
 
 function App() {
