+++ frontend-orange/src/components/tela-cadastro/CadastroForms.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 import { PropTypes } from 'prop-types';
 import { Button, TextField, Grid, Alert, Box, IconButton, InputAdornment } from '@mui/material';
 import { Visibility, VisibilityOff } from '@material-ui/icons';
-import styles from '../styles.jsx';
+import styles from '../../styles.jsx';
 
 const CadastroFormulario = ({ onCadastro }) => {
   const [form, setForm] = useState({
